Add oscillator module to project
The oscillator will be used to generate the signals in the synth.
Currently this only supports simple square waves, but this will allow a
simple hardware test to be constructed to prove out this method of
synthesis.  If it works, further waveforms and potentially other
synthesis methods may be introduced in the future.