Add initial envelope implementation
This includes a few of the basic envelope features that will be needed
for the synth, but is far from a complete implementation.  There are
some good ideas here that I would like to preserve in the version
history, but I am planning to start fresh in an upcomming commit that
will remove this and other existing code.