Add new module for synthesis engine
A new app subdirectory has been created in the slave project with a new
module, "synth".  This will contain the synthesis engine.  Currently it
just sends a ~50 Hz sine wave to the DAC.

A new python script, gen_sine_table.py, has been added to generate a
sine table for use in the synthesis module.