Add MPLAB project files
Debugging on the command line is pretty painful using MDB, so MPLAB will
have to be used for debugging.  These project files make it quick to get
the debugger running in MPLAB using the existing CMake build system.
There's a bit of hackery going in on the project Makefiles to bypass
MPLAB's own build system, but this seems to be the only way to use CMake
for the build while also having a decent debugging environment.