Add unit test framework for assembly
Some of the code on the slave core is written in assembly, so it can't
be tested with Ceedling on the host machine.  Instead, it has to be
built into a custom test binary to run on the target.  Some new CMake
targets have been added to facilitate this, and the slave project in
MPLAB has been updated with this new configuration.