Revert assembly unit test commits
While testing is important, this method was just way too much work.
It's easier to just manually test the code by stepping through with a
debugger.  The build system has to be extremely complex to manage the
separate master and slave binaries, and this is made worse by the lack
of simulator support for this chip and general MPLAB woes.  This may be
implemented eventually, but for now it doesn't seem to be worth it for
the tiny amount of code that it would test.