Add envelope generator module
The envelope generator will be used to control the gain of each
operator.  This is a fairly generic implementation, so it may be
possible to use it for other functions in the future.  It has not been
tested on actual hardware yet, but it passes all unit tests.  The next
step will be to add it to integrate it with the synthesis engine and do
some performance tests.