Add initial voice implementation
A new voice module has been added.  The voice is a single instance of
all of the features needed to play a single note.  Multiple voices can
be instantiated to play multiple notes simultaneously.  Currently, the
voice is implemented in pure C, which will be too slow.  Future commits
will add assembly for some of the critical DSP features.