Move to six voice polyphony
The synth engine is now capable of producing six voice polyphony, which
allows up to six notes to be played at the same time.  This is currently
implemented with a very simple round robin method, but this will
eventually be replaced with a least recently used algorithm in a future
commit.  The master core application and the audio module in the slave
core are still mostly spaghetti code, but it has allowed for quick
experimentation.  This nasty code will be replaced when more design
decisions have been made.