Add new messages module
This module is still very incomplete.  It will be responsible for
handling the incoming command messages and sending responses.  Some
basic error messages and a device info message have been implemented.

The linker scripts have been modified to increase the bootloader space
(the bootloader is now too big to fit on a single page), and also to
support reading the application version number from a standard location.

More error handling is still needed, as well as the erase/write/verify
messages.