Add bootloader app launch sequence handling
The run command has finally been implemented, and additional logic has
been added to main.c to control the application flow.  The first button
on the dev kit (RE7) is used to hold the device in bootloader mode if
held on reset.  This hasn't yet been tested with a real application -
some additional Python scripts will be needed to parse a hex file and
load real application code.